Fixed lifetime issue with higher-ranked trait bounds
After going through parse_entity - which ends by building the entity, supposedly getting rid of the EntityBuilder, the EntityBuilder was kept borrowed until the Script went out of scope. Turns out, this was a problem with the lifetimes I was using - so I summoned the magic of higher-ranked trait bounds to  assist.