Various design changes to get proper depth sort
No more "meshes" until I fully flesh out the design of the renderer. Each
Sprite has its own vertex and index buffer.

Upon construction, sprite components, instead of asking for what mesh to
use, instead ask for the bounds and uvs of a rectangle then construct the
vertices and indices of that rectangle in the render system.

To achieve y-depth sorting with tile maps, I opted to subdivide the tile
maps into "strips" (a row of 10 tiles, for now) instead of chunks (5x5
areas of tiles)

RenderIds allow me identify multiple types of renders, not just an entity's
component but the data on said component too (the tile map strip's index).