Conversion from the parsed tile map to the tile map component
Parsed layers, instead of holding individual tiles, now hold strips of them (10 currently, subject to change). This made conversion a helluva a lot easier.

Subsequently a chunk's position and dimensions fields are relative to the grid of strips, rather than the grid of tiles.

The render system now initializes and updates the tile map strips' vertex and index buffers (we flag that a specific strip needs updating by setting their buffers to None). This is because the system has access to the queue, unlike the tile map when we load its parsed data.

The tile map components now holds how many sub-textures fit in each dimension of its texture. This allows us to use the sub-texture indices (the elements of the strips field) from the parsed tile map.