Delegated various jobs of the render system
Instead of the render system updating and maintaining all kinds of render
components; I've given that job to individual systems like the tilemap and
sprite systems - which keep their respective components updated.

The render system now simply sorts the renders and draws them. Of course,
there needs a way to know what we are rendering and sorting - which
requires the knowledge of when a component is updated, inserted, or
removed (removal needs more work). In this case, the individual systems
mentioned above push their new renders in a collection stored in a specs
resource. This allows the render system to get this resource and
continually sort and render it every frame.